+++ R/zzz.R
@@ -3,11 +3,16 @@ utils::globalVariables("widget")
 
 .onLoad <- function(libname, pkgname) {
     op <- options()
-    toset <- !(c("RserveTS.format", "RserveTS.prettier_cmd") %in% names(op))
+    toset <- !(c(
+        "RserveTS.format",
+        "RserveTS.prettier_cmd",
+        "RserveTS.compile_dir"
+    ) %in% names(op))
     if (any(toset)) {
         options(list(
             RserveTS.format = FALSE,
-            RserveTS.prettier_cmd = NULL
+            RserveTS.prettier_cmd = NULL,
+            RserveTS.compile_dir = NULL
         )[toset])
     }
     invisible()